Former-commit-id: d6f3866812a11e5f29cb7493226994b29ba3dc57

+++ backend/src/main/java/com/ssafy/api/model/Friendship.java
@@ -26,7 +26,7 @@ public class Friendship {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "ID")
-	private int ID;
+	private long ID;
 
 	@Column(name = "FRIEND1_ID")
 	private long friend1ID;
